niScript: Do not store hstring is global variables since those have undefined initialization order ;
We keep the default macro which uses inline and a static variable this
is guaranteed to be thread safe and initialized in a deterministic
order. The linker will fold the duplicate definitions into one. cf
https://public.3.basecamp.com/p/UtY9bpCwKwjBmyYZ8AnBcWrr

The previous approach crashes at global variable init time (that is
program startup before main) on Windows likely because the MSVC linker
initialize the global hstring in a different order and the global _funcz
array referecing them get a null object instead of the initalized
hstring objects.